# openapi/spec.yaml
openapi: 3.0.0
info:
  title: SupALLM API
  version: "0.0.1"

# Apply security globally
security:
  - BearerAuth: []
paths:
...
  /projects/{projectId}/credentials:
    get:
      summary: List all credentials for a project
      operationId: listCredentials
      tags:
        - Credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: List of credentials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credential"
        "404":
          description: Project not found
    post:
      summary: "Create a credential for a project"
      operationId: createCredential
      tags:
        - Credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialRequest"
      responses:
        "201":
          description: "Credential created"
        "400":
          description: Bad request
        "404":
          description: Project not found
  /projects/{projectId}/credentials/{credentialId}:
    get:
      summary: "Get a credential by ID"
      operationId: getCredential
      tags:
        - Credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: credentialId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: "Credentials"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credential"
        "404":
          description: Credential or project not found
    patch:
      summary: "Update a credential"
      operationId: updateCredential
      tags:
        - Credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: credentialId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCredentialRequest"
      responses:
        "200":
          description: "Credential updated"
        "400":
          description: Bad request
        "404":
          description: Credential or project not found
    delete:
      summary: "Delete a credential"
      operationId: deleteCredential
      tags:
        - Credential
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: credentialId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: "Credential deleted"
        "404":
          description: Credential or project not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID:
      type: string
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
...
    ProviderType:
      type: string
      enum:
        - openai
        - anthropic
...
    Credential:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        provider:
          $ref: "#/components/schemas/ProviderType"
        apiKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - provider
        - apiKey
        - createdAt
        - updatedAt
    CreateCredentialRequest:
      type: object
      properties:
        name:
          type: string
        provider:
          $ref: "#/components/schemas/ProviderType"
        apiKey:
          type: string
      required:
        - name
        - provider
        - apiKey
    UpdateCredentialRequest:
      type: object
      properties:
        name:
          type: string
        apiKey:
          type: string
      required:
        - name
        - apiKey
...